hopefully fixed duplicate files
