--- frontend/.browserslistrc
@@ -1,2 +1,0 @@
-> 1%
-last 2 versions
