--- frontend/public/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <script src="https://kit.fontawesome.com/0e74e3bd2d.js"></script>
-    <title>CrudMongoLibraryjs</title>
-  </head>
-  <body>
-    <noscript>
-      <strong>We're sorry but crudmongolibraryjs doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
-    </noscript>
-    <div id="app"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
