--- frontend/src/assets/global.css
@@ -1,30 +1,0 @@
-
-@font-face {
-  font-family: 'Work Sans';
-  src: url('/assets/WorkSans-Regular');
-}
-
-.fade-leave, .fade-enter-to {
-  opacity: 1 !important;
-  transition-duration: .5s;
-}
-
-.fade-leave-to, .fade-enter {
-  opacity: 0 !important;
-  transition-duration: .5s;
-}
-
-body {
-  margin: 0;
-  font-family: 'Work Sans';
-}
-
-body, input {
-  color: #707070;
-}
-
-.card {
-  border-radius: 6px;
-  background-color: #E6EAED;
-  color: #808080;
-}
