--- frontend/tslint.json
@@ -1,36 +1,0 @@
-{
-  "defaultSeverity": "warning",
-  "extends": [
-    "tslint:recommended"
-  ],
-  "linterOptions": {
-    "exclude": [
-      "node_modules/**"
-    ]
-  },
-  "rules": {
-    "quotemark": [true, "single"],
-    "indent": [true, "spaces", 2],
-    "interface-name": false,
-    "ordered-imports": false,
-    "triple-equals": true,
-    "no-bitwise": false,
-    "only-arrow-functions": true,
-    "prefer-for-of": true,
-    "curly": [true, "as-needed"],
-    "no-conditional-assignment": true,
-    "no-namespace": false,
-    "no-string-literal": false,
-    "no-var-keyword": true,
-    "interface-over-type-literal": false,
-    "deprecation": true,
-    "no-require-imports": true,
-    "arrow-return-shorthand": true,
-    "arrow-parens": [true, "ban-single-arg-parens"],
-    "eofline": true,
-    "member-access": false,
-    "semicolon": [true, "never"],
-    "object-literal-sort-keys": false,
-    "no-consecutive-blank-lines": false
-  }
-}
