+++ public/index.html
@@ -4,7 +4,8 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>crudmongolibraryjs</title>
+    <script src="https://kit.fontawesome.com/0e74e3bd2d.js"></script>
+    <title>CrudMongoLibraryjs</title>
   </head>
   <body>
     <noscript>
