+++ tslint.json
@@ -14,6 +14,7 @@
     "interface-name": false,
     "ordered-imports": false,
     "triple-equals": true,
+    "no-bitwise": false,
     "only-arrow-functions": true,
     "prefer-for-of": true,
     "curly": [true, "as-needed"],
